+++ CHANGELOG.md
@@ -16,6 +16,12 @@ The following changes are pending, and will be applied on the next major release
 
 ## Unreleased
 
+### New feature
+
+- `denyAccessRequest` supports a new `customFields` option to add custom fields to the
+  resulting Access Denial. This option behaves the same as the `customFields` option
+  of `approveAccessRequest`.
+
 ### Patch changes
 
 - When issuing an Access Grant from an Access Request using `approveAccessRequest`,